KataGo 功能更新（未完成）

+++ katago/src/main/java/com/goplatform/katago/pojo/KataCount.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class KataCount {
 
-    private int black;
+    private double black;
 
-    private int white;
+    private double white;
 
 }
